requirements.txt
streamlit>=1.37
ccxt>=4.3
pandas>=2.0
python-dotenv>=1.0
plotly>=5.18
requests>=2.31
streamlit-autorefresh>=0.0.2
